animal_chess_bot/requirements.txt
pyautogui>=0.9.53
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
pywin32>=300
pynput>=1.7.0
